# Use `uv` for blackbox venv if available