It's faster than plain python -m pip

Before:

    ./blackbox/bootstrap.sh  8.30s user 1.03s system 72% cpu 12.850 total

After:

    ./blackbox/bootstrap.sh  0.07s user 0.60s system 120% cpu 0.557 total